requirements.txt
# Deep Agents & LangGraph
deepagents>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Web Crawling (for finding and extracting scholarly papers)
crawl4ai>=0.7.0
beautifulsoup4>=4.12.0

# Vector Database
qdrant-client>=1.12.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# PDF Processing
pypdf>=5.0.0
pymupdf>=1.24.0
pdf2image>=1.17.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.10.0
aiofiles>=24.1.0

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Monitoring and Utilities
python-dotenv>=1.0.0
psutil>=6.1.0
numpy>=1.26.0
tenacity>=9.0.0

# Logging and Debugging
colorlog>=6.9.0
structlog>=24.4.0
rich>=13.9.0
